refactor: 永続データ構造を persistent_ds/ に分離
永続・部分永続版は「永続かどうか」が名前を見た時点で分かるため、
専用ディレクトリにしても配置を思い出すコストが発生しない。
data_structure/ に混ぜる意味がないので分離する。

union_find の永続版 2 つ（persistent / partially_persistent）も
persistent_ds/ に含め、「名前に persistent が入るものはすべて
persistent_ds/」という機械的に判定できる規則にした。

data_structure/ は 33 → 25 ファイル。README・CLAUDE.md の規則も更新。

全 247 テストの -fsyntax-only、persistent_ds/ 全ヘッダの単体コンパイル、
mise run docs-check を確認。
