+++ CLAUDE.md
@@ -10,7 +10,8 @@
 - include ルートは **`lib/`**。ヘッダ間は `lib/` からの相対パス（例: `#include "number_theory/modint.hpp"`）。
 - **データ構造は `lib/data_structure/` に置く**。バリアント数でディレクトリを昇格・降格させると
   使うたびに「専用ディレクトリがあるか」を思い出す必要が生じるので、**ファミリ専用ディレクトリを
-  新設しない**（既存の例外は規模の大きい `segtree/`・`heap/` のみ）。
+  新設しない**。分けてよいのは配置が見た目で判別でき思い出す必要のない軸だけで、既存の例外は
+  `segtree/`・`heap/`・`persistent_ds/`（永続・部分永続版は名前に `persistent` が入るものすべて）。
 
 ## ビルド・テスト
