+++ lib/persistent_ds/persistent_stack.hpp
