+++ test/yosupo/data_structure/persistent_range_affine_range_sum.test.cpp
@@ -2,8 +2,8 @@
 #include <iostream>
 #include <utility>
 #include <vector>
-#include "data_structure/persistent_segment_tree.hpp"
 #include "number_theory/modint.hpp"
+#include "persistent_ds/persistent_segment_tree.hpp"
 
 using Mint = modint998;
